+++ core/persist/PersistenceProvider.js
@@ -6,7 +6,7 @@
  */
 
 import {XH} from '@xh/hoist/core';
-import {LocalStorageProvider, PrefProvider, DashViewProvider} from '../';
+import {LocalStorageProvider, PrefProvider, DashViewProvider, ManualProvider} from '../';
 import {
     isUndefined,
     cloneDeep,
@@ -28,6 +28,8 @@ import {throwIf} from '@xh/hoist/utils/js';
  * @see PrefProvider - stores state in a predefined Hoist application Preference.
  * @see LocalStorageProvider - stores state in browser local storage under a configured key.
  * @see DashViewProvider - stores state with other Dashboard-specific state via a `DashViewModel`.
+ * @see ManualProvider - provides API for application and components to provide their own
+ *      storage mechanism.
  */
 export class PersistenceProvider {
 
@@ -47,6 +49,7 @@ export class PersistenceProvider {
             if (rest.prefKey) type = 'pref';
             if (rest.localStorageKey) type = 'localStorage';
             if (rest.dashViewModel) type = 'dashView';
+            if (rest.getData || rest.setData) type = 'manual';
         }
 
         switch (type) {
@@ -56,6 +59,8 @@ export class PersistenceProvider {
                 return new LocalStorageProvider(rest);
             case `dashView`:
                 return new DashViewProvider(rest);
+            case 'manual':
+                return new ManualProvider(rest);
             default:
                 throw XH.exception(`Unknown Persistence Provider for type: ${type}`);
         }